fix: fixed access to users endpoints

+++ src/users/controllers/users.controller.ts
@@ -1,6 +1,7 @@
 import {
   Controller,
   Delete,
+  ForbiddenException,
   Get,
   Param,
   Post,
@@ -10,6 +11,8 @@ import { UsersService } from '../services/users.service';
 import { AuthGuard } from '@nestjs/passport';
 import { Roles } from '../decorators/roles.decorator';
 import { RoleGuard } from '../guards/role.guard';
+import { GetUser } from '../decorators/get-user.decorator';
+import { User } from '../entities/user.entity';
 
 @Controller('users')
 @UseGuards(AuthGuard())
@@ -24,17 +27,29 @@ export class UsersController {
   }
 
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.usersService.findOne(id);
+  findOne(@Param('id') id: string, @GetUser() user: User) {
+    if (user.id !== id && user.role !== 'admin') {
+      throw new ForbiddenException();
+    } else {
+      return this.usersService.findOne(id);
+    }
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.usersService.remove(id);
+  remove(@Param('id') id: string, @GetUser() user: User) {
+    if (user.id !== id && user.role !== 'admin') {
+      throw new ForbiddenException();
+    } else {
+      return this.usersService.remove(id);
+    }
   }
 
   @Post(':id/api-key')
-  generateApiKey(@Param('id') id: string) {
-    return this.usersService.generateApiKey(id);
+  generateApiKey(@Param('id') id: string, @GetUser() user: User) {
+    if (user.id !== id && user.role !== 'admin') {
+      throw new ForbiddenException();
+    } else {
+      return this.usersService.generateApiKey(id);
+    }
   }
 }
